integrate fixed string bsi
