+++ pkg/vm/engine/aoe/storage/layout/index/blkzonemap.go
@@ -17,7 +17,6 @@ package index
 import (
 	"bytes"
 	"github.com/matrixorigin/matrixone/pkg/container/types"
-	"github.com/matrixorigin/matrixone/pkg/container/vector"
 	"github.com/matrixorigin/matrixone/pkg/encoding"
 	buf "github.com/matrixorigin/matrixone/pkg/vm/engine/aoe/storage/buffer"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/aoe/storage/common"
@@ -500,274 +499,6 @@ func (i *BlockZoneMapIndex) Ge(v interface{}) bool {
 
 func (i *BlockZoneMapIndex) Btw(minv interface{}, maxv interface{}) bool {
 	return i.Le(minv) && i.Ge(maxv)
-}
-
-func BuildBlockZoneMapIndex(data *vector.Vector, t types.Type, colIdx int16, isSorted bool) (Index, error) {
-	switch t.Oid {
-	case types.T_int8:
-		vec := data.Col.([]int8)
-		var min, max int8
-		if isSorted {
-			min = vec[0]
-			max = vec[len(vec)-1]
-		} else {
-			min = vec[0]
-			max = min
-			for _, e := range vec {
-				if e > max {
-					max = e
-				}
-				if e < min {
-					min = e
-				}
-			}
-		}
-		zmi := NewBlockZoneMap(t, min, max, colIdx)
-		return zmi, nil
-	case types.T_int16:
-		vec := data.Col.([]int16)
-		var min, max int16
-		if isSorted {
-			min = vec[0]
-			max = vec[len(vec)-1]
-		} else {
-			min = vec[0]
-			max = min
-			for _, e := range vec {
-				if e > max {
-					max = e
-				}
-				if e < min {
-					min = e
-				}
-			}
-		}
-		zmi := NewBlockZoneMap(t, min, max, colIdx)
-		return zmi, nil
-	case types.T_int32:
-		vec := data.Col.([]int32)
-		var min, max int32
-		if isSorted {
-			min = vec[0]
-			max = vec[len(vec)-1]
-		} else {
-			min = vec[0]
-			max = min
-			for _, e := range vec {
-				if e > max {
-					max = e
-				}
-				if e < min {
-					min = e
-				}
-			}
-		}
-		zmi := NewBlockZoneMap(t, min, max, colIdx)
-		return zmi, nil
-	case types.T_int64:
-		vec := data.Col.([]int64)
-		var min, max int64
-		if isSorted {
-			min = vec[0]
-			max = vec[len(vec)-1]
-		} else {
-			min = vec[0]
-			max = min
-			for _, e := range vec {
-				if e > max {
-					max = e
-				}
-				if e < min {
-					min = e
-				}
-			}
-		}
-		zmi := NewBlockZoneMap(t, min, max, colIdx)
-		return zmi, nil
-	case types.T_uint8:
-		vec := data.Col.([]uint8)
-		var min, max uint8
-		if isSorted {
-			min = vec[0]
-			max = vec[len(vec)-1]
-		} else {
-			min = vec[0]
-			max = min
-			for _, e := range vec {
-				if e > max {
-					max = e
-				}
-				if e < min {
-					min = e
-				}
-			}
-		}
-		zmi := NewBlockZoneMap(t, min, max, colIdx)
-		return zmi, nil
-	case types.T_uint16:
-		vec := data.Col.([]uint16)
-		var min, max uint16
-		if isSorted {
-			min = vec[0]
-			max = vec[len(vec)-1]
-		} else {
-			min = vec[0]
-			max = min
-			for _, e := range vec {
-				if e > max {
-					max = e
-				}
-				if e < min {
-					min = e
-				}
-			}
-		}
-		zmi := NewBlockZoneMap(t, min, max, colIdx)
-		return zmi, nil
-	case types.T_uint32:
-		vec := data.Col.([]uint32)
-		var min, max uint32
-		if isSorted {
-			min = vec[0]
-			max = vec[len(vec)-1]
-		} else {
-			min = vec[0]
-			max = min
-			for _, e := range vec {
-				if e > max {
-					max = e
-				}
-				if e < min {
-					min = e
-				}
-			}
-		}
-		zmi := NewBlockZoneMap(t, min, max, colIdx)
-		return zmi, nil
-	case types.T_uint64:
-		vec := data.Col.([]uint64)
-		var min, max uint64
-		if isSorted {
-			min = vec[0]
-			max = vec[len(vec)-1]
-		} else {
-			min = vec[0]
-			max = min
-			for _, e := range vec {
-				if e > max {
-					max = e
-				}
-				if e < min {
-					min = e
-				}
-			}
-		}
-		zmi := NewBlockZoneMap(t, min, max, colIdx)
-		return zmi, nil
-	case types.T_float32:
-		vec := data.Col.([]float32)
-		var min, max float32
-		if isSorted {
-			min = vec[0]
-			max = vec[len(vec)-1]
-		} else {
-			min = vec[0]
-			max = min
-			for _, e := range vec {
-				if e > max {
-					max = e
-				}
-				if e < min {
-					min = e
-				}
-			}
-		}
-		zmi := NewBlockZoneMap(t, min, max, colIdx)
-		return zmi, nil
-	case types.T_float64:
-		vec := data.Col.([]float64)
-		var min, max float64
-		if isSorted {
-			min = vec[0]
-			max = vec[len(vec)-1]
-		} else {
-			min = vec[0]
-			max = min
-			for _, e := range vec {
-				if e > max {
-					max = e
-				}
-				if e < min {
-					min = e
-				}
-			}
-		}
-		zmi := NewBlockZoneMap(t, min, max, colIdx)
-		return zmi, nil
-	case types.T_date:
-		vec := data.Col.([]types.Date)
-		var min, max types.Date
-		if isSorted {
-			min = vec[0]
-			max = vec[len(vec)-1]
-		} else {
-			min = vec[0]
-			max = min
-			for _, e := range vec {
-				if e > max {
-					max = e
-				}
-				if e < min {
-					min = e
-				}
-			}
-		}
-		zmi := NewBlockZoneMap(t, min, max, colIdx)
-		return zmi, nil
-	case types.T_datetime:
-		vec := data.Col.([]types.Datetime)
-		var min, max types.Datetime
-		if isSorted {
-			min = vec[0]
-			max = vec[len(vec)-1]
-		} else {
-			min = vec[0]
-			max = min
-			for _, e := range vec {
-				if e > max {
-					max = e
-				}
-				if e < min {
-					min = e
-				}
-			}
-		}
-		zmi := NewBlockZoneMap(t, min, max, colIdx)
-		return zmi, nil
-	case types.T_char, types.T_varchar, types.T_json:
-		vec := data.Col.(*types.Bytes)
-		var min, max []byte
-		if isSorted {
-			min = vec.Get(0)
-			max = vec.Get(int64(len(vec.Lengths) - 1))
-		} else {
-			min = vec.Get(0)
-			max = min
-			for i := 0; i < len(vec.Lengths); i++ {
-				e := vec.Get(int64(i))
-				if bytes.Compare(e, max) > 0 {
-					max = e
-				}
-				if bytes.Compare(e, min) < 0 {
-					min = e
-				}
-			}
-		}
-		zmi := NewBlockZoneMap(t, min, max, colIdx)
-		return zmi, nil
-	default:
-		panic("unsupported")
-	}
 }
 
 func MockInt32ZmIndices(cols int) (indices []Index) {
