+++ pkg/vm/engine/aoe/storage/layout/index/numbsi.go
@@ -16,9 +16,7 @@ package index
 
 import (
 	"bytes"
-	"github.com/matrixorigin/matrixone/pkg/container/nulls"
 	"github.com/matrixorigin/matrixone/pkg/container/types"
-	"github.com/matrixorigin/matrixone/pkg/container/vector"
 	"github.com/matrixorigin/matrixone/pkg/encoding"
 	buf "github.com/matrixorigin/matrixone/pkg/vm/engine/aoe/storage/buffer"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/aoe/storage/common"
@@ -41,7 +39,16 @@ type NumericBsiIndex struct {
 	FreeFunc    buf.MemoryFreeFunc
 }
 
-func initNumericBsi(t types.Type, bitSize int) *bsi.NumericBSI {
+func NewNumericBsiIndex(t types.Type, bitSize int, colIdx int16) *NumericBsiIndex {
+	bsiIdx := getNumericBsi(t, bitSize)
+	return &NumericBsiIndex{
+		T:          t,
+		Col:        colIdx,
+		NumericBSI: *bsiIdx,
+	}
+}
+
+func getNumericBsi(t types.Type, bitSize int) *bsi.NumericBSI {
 	var bsiIdx bsi.BitSlicedIndex
 	switch t.Oid {
 	case types.T_int8, types.T_int16, types.T_int32, types.T_int64, types.T_date, types.T_datetime:
@@ -54,15 +61,6 @@ func initNumericBsi(t types.Type, bitSize int) *bsi.NumericBSI {
 		panic("not supported")
 	}
 	return bsiIdx.(*bsi.NumericBSI)
-}
-
-func NewNumericBsiIndex(t types.Type, bitSize int, colIdx int16) *NumericBsiIndex {
-	bsiIdx := initNumericBsi(t, bitSize)
-	return &NumericBsiIndex{
-		T:          t,
-		Col:        colIdx,
-		NumericBSI: *bsiIdx,
-	}
 }
 
 func NewNumericBsiEmptyNode(vf common.IVFile, useCompress bool, freeFunc buf.MemoryFreeFunc) buf.IMemoryNode {
@@ -231,213 +229,3 @@ func (i *NumericBsiIndex) Marshal() ([]byte, error) {
 	return bw.Bytes(), nil
 }
 
-func BuildNumericBsiIndex(data []*vector.Vector, t types.Type, colIdx int16, startPos int) (bsi.BitSlicedIndex, error) {
-	switch t.Oid {
-	case types.T_int8:
-		bsiIdx := NewNumericBsiIndex(t, 8, colIdx)
-		row := startPos
-		for _, part := range data {
-			column := part.Col.([]int8)
-			for _, val := range column {
-				if nulls.Contains(part.Nsp, uint64(row-startPos)) {
-					row++
-					continue
-				}
-				if err := bsiIdx.Set(uint64(row), val); err != nil {
-					return nil, err
-				}
-				row++
-			}
-		}
-		return bsiIdx, nil
-	case types.T_int16:
-		bsiIdx := NewNumericBsiIndex(t, 16, colIdx)
-		row := startPos
-		for _, part := range data {
-			column := part.Col.([]int16)
-			for _, val := range column {
-				if nulls.Contains(part.Nsp, uint64(row-startPos)) {
-					row++
-					continue
-				}
-				if err := bsiIdx.Set(uint64(row), val); err != nil {
-					return nil, err
-				}
-				row++
-			}
-		}
-		return bsiIdx, nil
-	case types.T_int32:
-		bsiIdx := NewNumericBsiIndex(t, 32, colIdx)
-		row := startPos
-		for _, part := range data {
-			column := part.Col.([]int32)
-			for _, val := range column {
-				if nulls.Contains(part.Nsp, uint64(row-startPos)) {
-					row++
-					continue
-				}
-				if err := bsiIdx.Set(uint64(row), val); err != nil {
-					return nil, err
-				}
-				row++
-			}
-		}
-		return bsiIdx, nil
-	case types.T_int64:
-		bsiIdx := NewNumericBsiIndex(t, 64, colIdx)
-		row := startPos
-		for _, part := range data {
-			column := part.Col.([]int64)
-			for _, val := range column {
-				if nulls.Contains(part.Nsp, uint64(row-startPos)) {
-					row++
-					continue
-				}
-				if err := bsiIdx.Set(uint64(row), val); err != nil {
-					return nil, err
-				}
-				row++
-			}
-		}
-		return bsiIdx, nil
-	case types.T_uint8:
-		bsiIdx := NewNumericBsiIndex(t, 8, colIdx)
-		row := startPos
-		for _, part := range data {
-			column := part.Col.([]uint8)
-			for _, val := range column {
-				if nulls.Contains(part.Nsp, uint64(row-startPos)) {
-					row++
-					continue
-				}
-				if err := bsiIdx.Set(uint64(row), val); err != nil {
-					return nil, err
-				}
-				row++
-			}
-		}
-		return bsiIdx, nil
-	case types.T_uint16:
-		bsiIdx := NewNumericBsiIndex(t, 16, colIdx)
-		row := startPos
-		for _, part := range data {
-			column := part.Col.([]uint16)
-			for _, val := range column {
-				if nulls.Contains(part.Nsp, uint64(row-startPos)) {
-					row++
-					continue
-				}
-				if err := bsiIdx.Set(uint64(row), val); err != nil {
-					return nil, err
-				}
-				row++
-			}
-		}
-		return bsiIdx, nil
-	case types.T_uint32:
-		bsiIdx := NewNumericBsiIndex(t, 32, colIdx)
-		row := startPos
-		for _, part := range data {
-			column := part.Col.([]uint32)
-			for _, val := range column {
-				if nulls.Contains(part.Nsp, uint64(row-startPos)) {
-					row++
-					continue
-				}
-				if err := bsiIdx.Set(uint64(row), val); err != nil {
-					return nil, err
-				}
-				row++
-			}
-		}
-		return bsiIdx, nil
-	case types.T_uint64:
-		bsiIdx := NewNumericBsiIndex(t, 64, colIdx)
-		row := startPos
-		for _, part := range data {
-			column := part.Col.([]uint64)
-			for _, val := range column {
-				if nulls.Contains(part.Nsp, uint64(row-startPos)) {
-					row++
-					continue
-				}
-				if err := bsiIdx.Set(uint64(row), val); err != nil {
-					return nil, err
-				}
-				row++
-			}
-		}
-		return bsiIdx, nil
-	case types.T_float32:
-		bsiIdx := NewNumericBsiIndex(t, 32, colIdx)
-		row := startPos
-		for _, part := range data {
-			column := part.Col.([]float32)
-			for _, val := range column {
-				if nulls.Contains(part.Nsp, uint64(row-startPos)) {
-					row++
-					continue
-				}
-				if err := bsiIdx.Set(uint64(row), val); err != nil {
-					return nil, err
-				}
-				row++
-			}
-		}
-		return bsiIdx, nil
-	case types.T_float64:
-		bsiIdx := NewNumericBsiIndex(t, 64, colIdx)
-		row := startPos
-		for _, part := range data {
-			column := part.Col.([]float64)
-			for _, val := range column {
-				if nulls.Contains(part.Nsp, uint64(row-startPos)) {
-					row++
-					continue
-				}
-				if err := bsiIdx.Set(uint64(row), val); err != nil {
-					return nil, err
-				}
-				row++
-			}
-		}
-		return bsiIdx, nil
-	case types.T_datetime:
-		bsiIdx := NewNumericBsiIndex(t, 64, colIdx)
-		row := startPos
-		for _, part := range data {
-			column := part.Col.([]types.Datetime)
-			for _, val := range column {
-				if nulls.Contains(part.Nsp, uint64(row-startPos)) {
-					row++
-					continue
-				}
-				if err := bsiIdx.Set(uint64(row), int64(val)); err != nil {
-					return nil, err
-				}
-				row++
-			}
-		}
-		return bsiIdx, nil
-	case types.T_date:
-		bsiIdx := NewNumericBsiIndex(t, 32, colIdx)
-		row := startPos
-		for _, part := range data {
-			column := part.Col.([]types.Date)
-			for _, val := range column {
-				if nulls.Contains(part.Nsp, uint64(row-startPos)) {
-					row++
-					continue
-				}
-				if err := bsiIdx.Set(uint64(row), int32(val)); err != nil {
-					return nil, err
-				}
-				row++
-			}
-		}
-		return bsiIdx, nil
-	default:
-		panic("unsupported type")
-	}
-}
