+++ pkg/vm/engine/aoe/storage/layout/index/segzonemap.go
@@ -17,7 +17,6 @@ package index
 import (
 	"bytes"
 	"github.com/matrixorigin/matrixone/pkg/container/types"
-	"github.com/matrixorigin/matrixone/pkg/container/vector"
 	"github.com/matrixorigin/matrixone/pkg/encoding"
 	buf "github.com/matrixorigin/matrixone/pkg/vm/engine/aoe/storage/buffer"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/aoe/storage/common"
@@ -721,623 +720,4 @@ func (i *SegmentZoneMapIndex) Ge(v interface{}) bool {
 
 func (i *SegmentZoneMapIndex) Btw(v interface{}) bool {
 	panic("TODO")
-}
-
-func BuildSegmentZoneMapIndex(data []*vector.Vector, t types.Type, colIdx int16, isSorted bool) (Index, error) {
-	switch t.Oid {
-	case types.T_int8:
-		var globalMin, globalMax int8
-		var partMins, partMaxs []interface{}
-		if isSorted {
-			for i, vec := range data {
-				part := vec.Col.([]int8)
-				partMins = append(partMins, part[0])
-				partMaxs = append(partMaxs, part[len(part) - 1])
-				if i == 0 {
-					globalMin = part[0]
-				}
-				if i == len(data) - 1 {
-					globalMax = part[len(part) - 1]
-				}
-			}
-		} else {
-			var minv, maxv int8
-			for i, vec := range data {
-				part := vec.Col.([]int8)
-				if i == 0 {
-					globalMin = part[i]
-					globalMax = part[i]
-				}
-				for j, v := range part {
-					if j == 0 {
-						minv = v
-						maxv = v
-					}
-					if v > globalMax {
-						globalMax = v
-					}
-					if v < globalMin {
-						globalMin = v
-					}
-					if v > maxv {
-						maxv = v
-					}
-					if v < minv {
-						minv = v
-					}
-				}
-				partMaxs = append(partMaxs, maxv)
-				partMins = append(partMins, minv)
-			}
-		}
-		zmi := NewSegmentZoneMap(t, globalMin, globalMax, colIdx, partMins, partMaxs)
-		return zmi, nil
-	case types.T_int16:
-		var globalMin, globalMax int16
-		var partMins, partMaxs []interface{}
-		if isSorted {
-			for i, vec := range data {
-				part := vec.Col.([]int16)
-				partMins = append(partMins, part[0])
-				partMaxs = append(partMaxs, part[len(part) - 1])
-				if i == 0 {
-					globalMin = part[0]
-				}
-				if i == len(data) - 1 {
-					globalMax = part[len(part) - 1]
-				}
-			}
-		} else {
-			var minv, maxv int16
-			for i, vec := range data {
-				part := vec.Col.([]int16)
-				if i == 0 {
-					globalMin = part[i]
-					globalMax = part[i]
-				}
-				for j, v := range part {
-					if j == 0 {
-						minv = v
-						maxv = v
-					}
-					if v > globalMax {
-						globalMax = v
-					}
-					if v < globalMin {
-						globalMin = v
-					}
-					if v > maxv {
-						maxv = v
-					}
-					if v < minv {
-						minv = v
-					}
-				}
-				partMaxs = append(partMaxs, maxv)
-				partMins = append(partMins, minv)
-			}
-		}
-		zmi := NewSegmentZoneMap(t, globalMin, globalMax, colIdx, partMins, partMaxs)
-		return zmi, nil
-	case types.T_int32:
-		var globalMin, globalMax int32
-		var partMins, partMaxs []interface{}
-		if isSorted {
-			for i, vec := range data {
-				part := vec.Col.([]int32)
-				partMins = append(partMins, part[0])
-				partMaxs = append(partMaxs, part[len(part) - 1])
-				if i == 0 {
-					globalMin = part[0]
-				}
-				if i == len(data) - 1 {
-					globalMax = part[len(part) - 1]
-				}
-			}
-		} else {
-			var minv, maxv int32
-			for i, vec := range data {
-				part := vec.Col.([]int32)
-				if i == 0 {
-					globalMin = part[i]
-					globalMax = part[i]
-				}
-				for j, v := range part {
-					if j == 0 {
-						minv = v
-						maxv = v
-					}
-					if v > globalMax {
-						globalMax = v
-					}
-					if v < globalMin {
-						globalMin = v
-					}
-					if v > maxv {
-						maxv = v
-					}
-					if v < minv {
-						minv = v
-					}
-				}
-				partMaxs = append(partMaxs, maxv)
-				partMins = append(partMins, minv)
-			}
-		}
-		zmi := NewSegmentZoneMap(t, globalMin, globalMax, colIdx, partMins, partMaxs)
-		return zmi, nil
-	case types.T_int64:
-		var globalMin, globalMax int64
-		var partMins, partMaxs []interface{}
-		if isSorted {
-			for i, vec := range data {
-				part := vec.Col.([]int64)
-				partMins = append(partMins, part[0])
-				partMaxs = append(partMaxs, part[len(part) - 1])
-				if i == 0 {
-					globalMin = part[0]
-				}
-				if i == len(data) - 1 {
-					globalMax = part[len(part) - 1]
-				}
-			}
-		} else {
-			var minv, maxv int64
-			for i, vec := range data {
-				part := vec.Col.([]int64)
-				if i == 0 {
-					globalMin = part[i]
-					globalMax = part[i]
-				}
-				for j, v := range part {
-					if j == 0 {
-						minv = v
-						maxv = v
-					}
-					if v > globalMax {
-						globalMax = v
-					}
-					if v < globalMin {
-						globalMin = v
-					}
-					if v > maxv {
-						maxv = v
-					}
-					if v < minv {
-						minv = v
-					}
-				}
-				partMaxs = append(partMaxs, maxv)
-				partMins = append(partMins, minv)
-			}
-		}
-		zmi := NewSegmentZoneMap(t, globalMin, globalMax, colIdx, partMins, partMaxs)
-		return zmi, nil
-	case types.T_uint8:
-		var globalMin, globalMax uint8
-		var partMins, partMaxs []interface{}
-		if isSorted {
-			for i, vec := range data {
-				part := vec.Col.([]uint8)
-				partMins = append(partMins, part[0])
-				partMaxs = append(partMaxs, part[len(part) - 1])
-				if i == 0 {
-					globalMin = part[0]
-				}
-				if i == len(data) - 1 {
-					globalMax = part[len(part) - 1]
-				}
-			}
-		} else {
-			var minv, maxv uint8
-			for i, vec := range data {
-				part := vec.Col.([]uint8)
-				if i == 0 {
-					globalMin = part[i]
-					globalMax = part[i]
-				}
-				for j, v := range part {
-					if j == 0 {
-						minv = v
-						maxv = v
-					}
-					if v > globalMax {
-						globalMax = v
-					}
-					if v < globalMin {
-						globalMin = v
-					}
-					if v > maxv {
-						maxv = v
-					}
-					if v < minv {
-						minv = v
-					}
-				}
-				partMaxs = append(partMaxs, maxv)
-				partMins = append(partMins, minv)
-			}
-		}
-		zmi := NewSegmentZoneMap(t, globalMin, globalMax, colIdx, partMins, partMaxs)
-		return zmi, nil
-	case types.T_uint16:
-		var globalMin, globalMax uint16
-		var partMins, partMaxs []interface{}
-		if isSorted {
-			for i, vec := range data {
-				part := vec.Col.([]uint16)
-				partMins = append(partMins, part[0])
-				partMaxs = append(partMaxs, part[len(part) - 1])
-				if i == 0 {
-					globalMin = part[0]
-				}
-				if i == len(data) - 1 {
-					globalMax = part[len(part) - 1]
-				}
-			}
-		} else {
-			var minv, maxv uint16
-			for i, vec := range data {
-				part := vec.Col.([]uint16)
-				if i == 0 {
-					globalMin = part[i]
-					globalMax = part[i]
-				}
-				for j, v := range part {
-					if j == 0 {
-						minv = v
-						maxv = v
-					}
-					if v > globalMax {
-						globalMax = v
-					}
-					if v < globalMin {
-						globalMin = v
-					}
-					if v > maxv {
-						maxv = v
-					}
-					if v < minv {
-						minv = v
-					}
-				}
-				partMaxs = append(partMaxs, maxv)
-				partMins = append(partMins, minv)
-			}
-		}
-		zmi := NewSegmentZoneMap(t, globalMin, globalMax, colIdx, partMins, partMaxs)
-		return zmi, nil
-	case types.T_uint32:
-		var globalMin, globalMax uint32
-		var partMins, partMaxs []interface{}
-		if isSorted {
-			for i, vec := range data {
-				part := vec.Col.([]uint32)
-				partMins = append(partMins, part[0])
-				partMaxs = append(partMaxs, part[len(part) - 1])
-				if i == 0 {
-					globalMin = part[0]
-				}
-				if i == len(data) - 1 {
-					globalMax = part[len(part) - 1]
-				}
-			}
-		} else {
-			var minv, maxv uint32
-			for i, vec := range data {
-				part := vec.Col.([]uint32)
-				if i == 0 {
-					globalMin = part[i]
-					globalMax = part[i]
-				}
-				for j, v := range part {
-					if j == 0 {
-						minv = v
-						maxv = v
-					}
-					if v > globalMax {
-						globalMax = v
-					}
-					if v < globalMin {
-						globalMin = v
-					}
-					if v > maxv {
-						maxv = v
-					}
-					if v < minv {
-						minv = v
-					}
-				}
-				partMaxs = append(partMaxs, maxv)
-				partMins = append(partMins, minv)
-			}
-		}
-		zmi := NewSegmentZoneMap(t, globalMin, globalMax, colIdx, partMins, partMaxs)
-		return zmi, nil
-	case types.T_uint64:
-		var globalMin, globalMax uint64
-		var partMins, partMaxs []interface{}
-		if isSorted {
-			for i, vec := range data {
-				part := vec.Col.([]uint64)
-				partMins = append(partMins, part[0])
-				partMaxs = append(partMaxs, part[len(part) - 1])
-				if i == 0 {
-					globalMin = part[0]
-				}
-				if i == len(data) - 1 {
-					globalMax = part[len(part) - 1]
-				}
-			}
-		} else {
-			var minv, maxv uint64
-			for i, vec := range data {
-				part := vec.Col.([]uint64)
-				if i == 0 {
-					globalMin = part[i]
-					globalMax = part[i]
-				}
-				for j, v := range part {
-					if j == 0 {
-						minv = v
-						maxv = v
-					}
-					if v > globalMax {
-						globalMax = v
-					}
-					if v < globalMin {
-						globalMin = v
-					}
-					if v > maxv {
-						maxv = v
-					}
-					if v < minv {
-						minv = v
-					}
-				}
-				partMaxs = append(partMaxs, maxv)
-				partMins = append(partMins, minv)
-			}
-		}
-		zmi := NewSegmentZoneMap(t, globalMin, globalMax, colIdx, partMins, partMaxs)
-		return zmi, nil
-	case types.T_float32:
-		var globalMin, globalMax float32
-		var partMins, partMaxs []interface{}
-		if isSorted {
-			for i, vec := range data {
-				part := vec.Col.([]float32)
-				partMins = append(partMins, part[0])
-				partMaxs = append(partMaxs, part[len(part) - 1])
-				if i == 0 {
-					globalMin = part[0]
-				}
-				if i == len(data) - 1 {
-					globalMax = part[len(part) - 1]
-				}
-			}
-		} else {
-			var minv, maxv float32
-			for i, vec := range data {
-				part := vec.Col.([]float32)
-				if i == 0 {
-					globalMin = part[i]
-					globalMax = part[i]
-				}
-				for j, v := range part {
-					if j == 0 {
-						minv = v
-						maxv = v
-					}
-					if v > globalMax {
-						globalMax = v
-					}
-					if v < globalMin {
-						globalMin = v
-					}
-					if v > maxv {
-						maxv = v
-					}
-					if v < minv {
-						minv = v
-					}
-				}
-				partMaxs = append(partMaxs, maxv)
-				partMins = append(partMins, minv)
-			}
-		}
-		zmi := NewSegmentZoneMap(t, globalMin, globalMax, colIdx, partMins, partMaxs)
-		return zmi, nil
-	case types.T_float64:
-		var globalMin, globalMax float64
-		var partMins, partMaxs []interface{}
-		if isSorted {
-			for i, vec := range data {
-				part := vec.Col.([]float64)
-				partMins = append(partMins, part[0])
-				partMaxs = append(partMaxs, part[len(part) - 1])
-				if i == 0 {
-					globalMin = part[0]
-				}
-				if i == len(data) - 1 {
-					globalMax = part[len(part) - 1]
-				}
-			}
-		} else {
-			var minv, maxv float64
-			for i, vec := range data {
-				part := vec.Col.([]float64)
-				if i == 0 {
-					globalMin = part[i]
-					globalMax = part[i]
-				}
-				for j, v := range part {
-					if j == 0 {
-						minv = v
-						maxv = v
-					}
-					if v > globalMax {
-						globalMax = v
-					}
-					if v < globalMin {
-						globalMin = v
-					}
-					if v > maxv {
-						maxv = v
-					}
-					if v < minv {
-						minv = v
-					}
-				}
-				partMaxs = append(partMaxs, maxv)
-				partMins = append(partMins, minv)
-			}
-		}
-		zmi := NewSegmentZoneMap(t, globalMin, globalMax, colIdx, partMins, partMaxs)
-		return zmi, nil
-	case types.T_datetime:
-		var globalMin, globalMax types.Datetime
-		var partMins, partMaxs []interface{}
-		if isSorted {
-			for i, vec := range data {
-				part := vec.Col.([]types.Datetime)
-				partMins = append(partMins, part[0])
-				partMaxs = append(partMaxs, part[len(part) - 1])
-				if i == 0 {
-					globalMin = part[0]
-				}
-				if i == len(data) - 1 {
-					globalMax = part[len(part) - 1]
-				}
-			}
-		} else {
-			var minv, maxv types.Datetime
-			for i, vec := range data {
-				part := vec.Col.([]types.Datetime)
-				if i == 0 {
-					globalMin = part[i]
-					globalMax = part[i]
-				}
-				for j, v := range part {
-					if j == 0 {
-						minv = v
-						maxv = v
-					}
-					if v > globalMax {
-						globalMax = v
-					}
-					if v < globalMin {
-						globalMin = v
-					}
-					if v > maxv {
-						maxv = v
-					}
-					if v < minv {
-						minv = v
-					}
-				}
-				partMaxs = append(partMaxs, maxv)
-				partMins = append(partMins, minv)
-			}
-		}
-		zmi := NewSegmentZoneMap(t, globalMin, globalMax, colIdx, partMins, partMaxs)
-		return zmi, nil
-	case types.T_date:
-		var globalMin, globalMax types.Date
-		var partMins, partMaxs []interface{}
-		if isSorted {
-			for i, vec := range data {
-				part := vec.Col.([]types.Date)
-				partMins = append(partMins, part[0])
-				partMaxs = append(partMaxs, part[len(part) - 1])
-				if i == 0 {
-					globalMin = part[0]
-				}
-				if i == len(data) - 1 {
-					globalMax = part[len(part) - 1]
-				}
-			}
-		} else {
-			var minv, maxv types.Date
-			for i, vec := range data {
-				part := vec.Col.([]types.Date)
-				if i == 0 {
-					globalMin = part[i]
-					globalMax = part[i]
-				}
-				for j, v := range part {
-					if j == 0 {
-						minv = v
-						maxv = v
-					}
-					if v > globalMax {
-						globalMax = v
-					}
-					if v < globalMin {
-						globalMin = v
-					}
-					if v > maxv {
-						maxv = v
-					}
-					if v < minv {
-						minv = v
-					}
-				}
-				partMaxs = append(partMaxs, maxv)
-				partMins = append(partMins, minv)
-			}
-		}
-		zmi := NewSegmentZoneMap(t, globalMin, globalMax, colIdx, partMins, partMaxs)
-		return zmi, nil
-	case types.T_char, types.T_varchar, types.T_json:
-		var globalMin, globalMax []byte
-		var partMins, partMaxs []interface{}
-		if isSorted {
-			for i, vec := range data {
-				part := vec.Col.(*types.Bytes)
-				partMins = append(partMins, part.Get(0))
-				partMaxs = append(partMaxs, part.Get(int64(len(part.Offsets)-1)))
-				if i == 0 {
-					globalMin = part.Get(0)
-				}
-				if i == len(data) - 1 {
-					globalMax = part.Get(int64(len(part.Offsets)-1))
-				}
-			}
-		} else {
-			var minv, maxv []byte
-			for i, vec := range data {
-				part := vec.Col.(*types.Bytes)
-				if i == 0 {
-					globalMin = part.Get(0)
-					globalMax = part.Get(0)
-				}
-				for j := int64(0); j < int64(len(part.Offsets)); j++ {
-					v := part.Get(j)
-					if j == 0 {
-						minv = v
-						maxv = v
-					}
-					if bytes.Compare(v, globalMax) > 0 {
-						globalMax = v
-					}
-					if bytes.Compare(v, globalMin) < 0 {
-						globalMin = v
-					}
-					if bytes.Compare(v, maxv) > 0 {
-						maxv = v
-					}
-					if bytes.Compare(v, minv) < 0 {
-						minv = v
-					}
-				}
-				partMaxs = append(partMaxs, maxv)
-				partMins = append(partMins, minv)
-			}
-		}
-		zmi := NewSegmentZoneMap(t, globalMin, globalMax, colIdx, partMins, partMaxs)
-		return zmi, nil
-	default:
-		panic("unsupported type")
-	}
 }
